commit after code refactoring

+++ routes/api.php
@@ -17,42 +17,43 @@ Route::post('/login','AuthenticateController@authenticate');
 Route::get('/login','AuthenticateController@login') -> name('login');
 
 //Route for customer
-Route::post('/createCustomer','FirmController@store');
-Route::patch('/updateCustomer/{id}','FirmController@update');
-Route::delete('/deleteCustomer/{id}','FirmController@destroy');
-Route::get('/getCustomer','FirmController@index');
-Route::get('/showCustomer/{id}','FirmController@show');
-Route::get('/listCustomer','FirmController@lists');
+Route::post('/createCustomer','FirmController@store')->middleware('myauth');
+Route::patch('/updateCustomer/{id}','FirmController@update')->middleware('myauth');
+Route::delete('/deleteCustomer/{id}','FirmController@destroy')->middleware('myauth');
+Route::get('/getCustomer','FirmController@index')->middleware('myauth');
+Route::get('/showCustomer/{id}','FirmController@show')->middleware('myauth');
+//This route is created to show drop down list on frontend 
+Route::get('/listCustomer','FirmController@lists')->middleware('myauth');
 
-//API related to bill creation viewing
-Route::delete('/deleteBill/{id}','BillController@destroy');
-Route::get('/getBill','BillController@index');
-Route::post('/createBill','BillController@store');
+//API related to bill creation viewing and deletion
+Route::get('/getBill','BillController@index')->middleware('myauth');
+Route::post('/createBill','BillController@store')->middleware('myauth');
+Route::delete('/deleteBill/{id}','BillController@destroy')->middleware('myauth');
 
 //API relted to reports
-Route::post('/betweenDate', 'ReportController@BetweenDates');
-Route::get('/getBill/{invoice}','ReportController@singleBill');
-Route::get('/invoiceNumber','ReportController@nextInvoice');
-Route::get('/weekSale','ReportController@weekSale');
+Route::post('/betweenDate', 'ReportController@BetweenDates')->middleware('myauth');
+Route::get('/getBill/{invoice}','ReportController@singleBill')->middleware('myauth');
+Route::get('/invoiceNumber','ReportController@nextInvoice')->middleware('myauth');
+Route::get('/weekSale','ReportController@weekSale')->middleware('myauth');
 
 //Routes for users OR admin related details
-Route::get('/userDetails','UserDetailController@show');
-Route::patch('/userDetails/{id}','UserDetailController@update');
-Route::delete('/userDetails/{id}','UserDetailController@destroy');
+Route::get('/userDetails','UserDetailController@show')->middleware('myauth');
+Route::patch('/userDetails/{id}','UserDetailController@update')->middleware('myauth');
+Route::delete('/userDetails/{id}','UserDetailController@destroy')->middleware('myauth');
 
 
 // Routes for states
-Route::get('/getStates','StateController@index');
-Route::post('/addState','StateController@store');
-Route::patch('/updateState/{state_code}','StateController@update');
-Route::delete('/deleteState/{state_code}', 'StateController@destroy');
+Route::get('/getStates','StateController@index')->middleware('myauth');
+Route::post('/addState','StateController@store')->middleware('myauth');
+Route::patch('/updateState/{state_code}','StateController@update')->middleware('myauth');
+Route::delete('/deleteState/{state_code}', 'StateController@destroy')->middleware('myauth');
 
 
 //Routes for products
-Route::get('/getProducts','ProductController@index');
-Route::post('/addProduct','ProductController@store');
-Route::patch('/updateProduct/{id}','ProductController@update');
-Route::delete('/deleteProduct/{id}', 'ProductController@destroy');
+Route::get('/getProducts','ProductController@index')->middleware('myauth');
+Route::post('/addProduct','ProductController@store')->middleware('myauth');
+Route::patch('/updateProduct/{id}','ProductController@update')->middleware('myauth');
+Route::delete('/deleteProduct/{id}', 'ProductController@destroy')->middleware('myauth');
 
 //Testing Route
 Route::get('/testing','ReportController@weekSale');
